Moved the CsrfProtector class into a separate file
It used to be in the validator module, but isn't technically a validator
(even though it does validation). It's middleware. So I've moved it into
a separate module.
